+++ components/src/core/drawer/drawer-nav-item.tsx
@@ -18,7 +18,7 @@ export type NavItem = {
     // IconClass is replaced by the 'icon' property.
     Omit<InfoListItemProps, 'IconClass'>;
 
-export type DrawerNavItemProps = Omit<Omit<InfoListItemProps, 'styles'>, 'title'> & {
+export type DrawerNavItemProps = {
     depth: number;
     expanded: boolean;
     expandHandler?: () => void;
@@ -28,6 +28,7 @@ export type DrawerNavItemProps = Omit<Omit<InfoListItemProps, 'styles'>, 'title'
     styles?: {
         root?: StyleProp<ViewStyle>;
         activeBackground?: StyleProp<ViewStyle>;
+        expandIcon?: StyleProp<ViewStyle>;
         infoListItem?: InfoListItemProps['styles'];
     };
     /** Overrides for theme */
@@ -49,11 +50,16 @@ const makeStyles = (props: DrawerNavItemProps): any =>
             borderBottomRightRadius: props.navItem.activeItemBackgroundShape === 'square' ? 0 : 24,
             opacity: 0.9,
         },
+        expandIcon: {
+            display: 'flex',
+            alignItems: 'center',
+            marginLeft: 16,
+        },
     });
 
 export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
     const defaultStyles = makeStyles(props);
-    const { depth, expanded, expandHandler, navGroupProps, navItem, styles = {}, ...infoListItemProps } = props;
+    const { depth, expanded, expandHandler, navGroupProps, navItem, styles = {} } = props;
     const theme = useTheme();
 
     const icon = !depth ? (navItem as NavItem).icon : undefined;
@@ -85,7 +91,16 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
                     <InfoListItem
                         dense
                         {...navItem}
-                        rightComponent={rightIcon}
+                        rightComponent={
+                            (navItem.rightComponent || rightIcon) && (
+                                <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+                                    {navItem.rightComponent}
+                                    {rightIcon && (
+                                        <View style={[defaultStyles.expandIcon, styles.expandIcon]}>{rightIcon}</View>
+                                    )}
+                                </View>
+                            )
+                        }
                         backgroundColor={'transparent'}
                         iconColor={active ? navItem.activeItemIconColor : navItem.iconColor || navItem.itemIconColor}
                         fontColor={active ? navItem.activeItemFontColor : navItem.fontColor || navItem.itemFontColor}
@@ -109,7 +124,6 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
                             ),
                             ...otherILI,
                         }}
-                        {...infoListItemProps}
                     />
                 </View>
             )}